entityTemplate: width/height max of 15

+++ entityTemplates.cpp
@@ -7,8 +7,8 @@ void loadEntity(Entity& entity, EntityType entityType) {
 
     uint8_t* start = entityProperties + (entityType * NUM_ENTITY_PROPS);
 
-    const uint8_t widthAndHeight = pgm_read_byte(start++);
-    const uint8_t healthAndDamage = pgm_read_byte(start++);
+    const uint8_t widthAndHeight = (uint8_t)pgm_read_byte(start++);
+    const uint8_t healthAndDamage = (uint8_t)pgm_read_byte(start++);
     entity.duration = pgm_read_byte(start++);
 
     entity.width = (widthAndHeight >> 4) + 1;
